Block ingestion app round 2

+++ src/bin/mainnet-test.rs
@@ -1,7 +1,5 @@
 use bytesize::ByteSize;
-
 use clap::Parser;
-
 use mina_indexer::{
     block::{parser::BlockParser, BlockHash},
     state::{ledger::genesis, IndexerState},
@@ -17,18 +15,23 @@ struct Args {
     blocks_dir: PathBuf,
     #[arg(short, long, default_value_t = 10_000)]
     max_block_count: u32,
+    #[arg(short, long, default_value_t = 5000)]
+    report_freq: u32,
+    #[arg(short, long, default_value_t = false)]
+    persist_db: bool,
+    #[arg(short, long, default_value_t = false)]
+    verbose: bool,
 }
 
 #[tokio::main]
 async fn main() {
     let args = Args::parse();
     let blocks_dir = args.blocks_dir;
+    let freq = args.report_freq;
 
     assert!(blocks_dir.is_dir(), "Should be a dir path");
 
     let max_block_count = args.max_block_count;
-
-    let freq = 5000;
 
     let mut bp = BlockParser::new(&blocks_dir).unwrap();
 
@@ -37,8 +40,13 @@ async fn main() {
 
     const GENESIS_HASH: &str = "3NKeMoncuHab5ScarV5ViyF16cJPT4taWNSaTLS64Dp67wuXigPZ";
     let genesis_path = &PathBuf::from("./tests/data/genesis_ledgers/mainnet.json");
-    let genesis_root = genesis::parse_file(genesis_path).await.unwrap();
 
+    let parse_genesis_time = Instant::now();
+    let genesis_root = genesis::parse_file(genesis_path).await.unwrap();
+    let parse_genesis_time = parse_genesis_time.elapsed();
+    println!("Genesis ledger parsing time: {parse_genesis_time:?}");
+
+    let total_time = Instant::now();
     let mut state = IndexerState::new(
         BlockHash(GENESIS_HASH.to_string()),
         genesis_root.ledger,
@@ -47,6 +55,8 @@ async fn main() {
         None,
     )
     .unwrap();
+    let sorting_time = total_time.elapsed();
+    println!("Sorting time: {sorting_time:?}\n");
 
     let mut max_branches = 1;
     let mut max_root_len = 0;
@@ -54,19 +64,29 @@ async fn main() {
     let mut max_dangling_len = 0;
     let mut max_dangling_height = 0;
     let mut block_count = 1;
-    let total = Instant::now();
-    let mut adding = Duration::new(0, 0);
+    let mut highest_seq_height = 2;
+
+    let mut floor_minutes = 0;
+    let mut adding_time = Duration::new(0, 0);
+    let mut parsing_time = Duration::new(0, 0);
 
     for _ in 1..max_block_count {
+        // Report every passing minute
+        if args.verbose && total_time.elapsed().as_secs() % 60 > floor_minutes {
+            println!("Time elapsed: {:?}", total_time.elapsed());
+            floor_minutes += 1;
+        }
+
+        // Report every freq blocks
         if block_count % freq == 0 {
             println!("=== Progress #{} ===", block_count / freq);
             println!("Blocks:  {block_count}");
-            println!("Total:   {:?}", total.elapsed());
+            println!("Total:   {:?}", total_time.elapsed());
 
-            let blocks_per_sec = block_count as f64 / adding.as_secs_f64();
+            let blocks_per_sec = block_count as f64 / adding_time.as_secs_f64();
             println!("\n~~~ Add to state ~~~");
-            println!("Avg:     {:?}", adding / block_count);
-            println!("Total:   {adding:?}");
+            println!("Avg:     {:?}", adding_time / block_count);
+            println!("Total:   {adding_time:?}");
             println!("Per sec: {blocks_per_sec:?} blocks");
             println!("Per hr:  {:?} blocks", blocks_per_sec * 3600.);
 
@@ -74,10 +94,11 @@ async fn main() {
             println!("Max num:             {max_branches}");
             println!("Max root length:     {max_root_len}");
             println!("Max root height:     {max_root_height}");
-            println!("Max dangling len:    {max_dangling_len}");
+            println!("Max dangling length: {max_dangling_len}");
             println!("Max dangling height: {max_dangling_height}\n");
         }
 
+        let parse_time = Instant::now();
         match bp.next().await {
             Err(err) => {
                 println!("{err:?}");
@@ -87,9 +108,35 @@ async fn main() {
                 break;
             }
             Ok(Some(block)) => {
+                parsing_time += parse_time.elapsed();
+
                 let add = Instant::now();
                 state.add_block(&block).unwrap();
-                adding += add.elapsed();
+                adding_time += add.elapsed();
+
+                match block.blockchain_length {
+                    None => println!("Block with no height! state_hash: {:?}", block.state_hash),
+                    Some(n) => match n.cmp(&highest_seq_height) {
+                        std::cmp::Ordering::Less => {
+                            println!(
+                                "Another block of height: {n}! state_hash: {:?}",
+                                block.state_hash
+                            );
+                        }
+                        std::cmp::Ordering::Equal => highest_seq_height += 1,
+                        std::cmp::Ordering::Greater => {
+                            println!("Expected {highest_seq_height}, instead got height {n}, state_hash: {:?}", block.state_hash);
+                        }
+                    },
+                }
+
+                if args.verbose {
+                    println!(
+                        "Added block (length: {}, state_hash: {:?})",
+                        block.blockchain_length.unwrap_or(0),
+                        block.state_hash
+                    );
+                }
 
                 // update metrics
                 block_count += 1;
@@ -104,14 +151,19 @@ async fn main() {
         }
     }
 
-    let total_add = adding;
-    let total_time = total.elapsed();
+    let total_add = adding_time;
+    let total_time = total_time.elapsed();
 
     println!("\n~~~~~~~~~~~~~~~~~~");
     println!("~~~ Benchmarks ~~~");
     println!("~~~~~~~~~~~~~~~~~~");
+    println!("Sorting: {sorting_time:?}");
     println!("Blocks:  {block_count}");
     println!("Total:   {total_time:?}");
+
+    println!("~~~ Parsing ~~~");
+    println!("Genesis ledger: {parse_genesis_time:?}");
+    println!("Blocks:         {parsing_time:?}");
 
     let blocks_per_sec = block_count as f64 / total_add.as_secs_f64();
     println!("\n~~~ Add to state ~~~");
@@ -124,7 +176,7 @@ async fn main() {
     println!("Max num:             {max_branches}");
     println!("Root height:         {}", &state.root_branch.height());
     println!("Root length:         {}", &state.root_branch.len());
-    println!("Max dangling len:    {max_dangling_len}");
+    println!("Max dangling length: {max_dangling_len}");
     println!("Max dangling height: {max_dangling_height}\n");
 
     println!("Estimated time to ingest all (~640_000) mainnet blocks at this rate:");
@@ -164,5 +216,7 @@ async fn main() {
     );
     println!("{}", state.block_store.as_ref().unwrap().db_stats());
 
-    tokio::fs::remove_dir_all(store_dir).await.unwrap();
+    if !args.persist_db {
+        tokio::fs::remove_dir_all(store_dir).await.unwrap();
+    }
 }
